requirements.txt
django==2.1.1
pytz==2018.5
djangorestframework==3.8.2
certifi==2018.8.24
chardet==3.0.4
idna==2.7
requests==2.19.1
urllib3==1.23
amqp=2.3.2
billiard==3.5.0.4
celery==4.2.1
kombu==4.2.1
vine==1.1.4
